Added test summary and less verbose output
The test now ends with a summary of the number of passed/failed tests. This also
lets us not output a line for each and every test, particularly when they
passed, since there isn't much info there, but we still want to make sure that
the expected number of tests did run.

We now have a way to control verbosity with `VERBOSE` env var, which defaults to
0, and silences individual test outputs, unless they failed.