Use `ed` to prepend text to a file.
When updating a file in-place to prepend the license text, rather than creating
a new temporary file and overwriting the old file (which may lose permissions as
well as owner and group settings), use `ed` (the standard editor) to prepend the
license text to the existing file.

This fixes issue #6 and is a more elegant solution than manually copying over
the permissions, owner, group (which can be done via `stat -f '%p'`), because
those commands may also fail, if the user does not have the appropriate level of
access to modify file permissions or ownership.