Handle symlink chains properly.
Added a cross-platform [1] method of resolving a potential chain of symlinks, if
`autogen.sh` is not referred to directly as a file but as a symlink, which in
tern may point to another symlink, etc.

Different operating systems provide various versions of `realpath`, `readlink`,
`greadlink`, etc., and even if the same tool is available on multiple platforms,
it may vary in feature support and available flags, which complicates the
resolution of symlinks.

This addresses issue #3.

[1] Tested on Linux and OS X.