Make separator comment line optional.
This is a backwards-incompatible change: the default behavior is now to leave a
blank line between the header license comment and the top-level file comment,
instead of a line of comment characters.

This change provides the following benefits:

* the output is much cleaner to read
* editors can automatically collapse the license comments
* the output should be compliant with eslint

To get the original behavior, this change provides the (optional) --separator
flag, whose output matches the previous behavior. Existing tests were updated to
match the new behavior; additional tests were added to demonstrate the behavior
using the --separator flag.

Fixes issue #12 and #16.