Move Linux-specific config into Linux builds.
Removed APT config from the top-level scope since it does not apply to macOS X,
but only to the two Linux builds.

Although this required duplication of the APT config for `bc` and `ed` package
installation, this makes it much clearer as to what needs to happen for us to be
able to drop the `sudo` requirement for each of our Linux builds, and enable us
to use the faster, more efficient container-based infrastructure.