Remove `ex` and `ed` methods of file editing.
We can just use `cat` which is simpler, easier to understand, always available,
does not require installing a separate package or a config script, and uses a
temporary file just like the other solutions.

Added an explicit cleanup of the temporary file in addition to `trap` because
future calls to `trap` overwrite previous calls, so we may end up leaving
temporary files around if we don't clean up manually.