Add flag to skip the top-level "TODO" comment
Add `--top-level-comment` and the alias `--tlc` to enable the feature (which is
the default), as well as similar flags `--no-top-level-comment` and `--no-tlc`
to disable the functionality.

This enables users to avoid the "TODO" comment when they don't want to have it,
as well as to the mass-update existing files to add the license comment headers
without also adding the "TODO" comment for those files which may already have
those comments, or prefer to have none added at all, since they don't want to
have to manually update each and every one of those files.