Moved `testdata` under the `tests` dir.
No need to have `testdata` be a top-level directory.

Also, added missing entry for `longflags_test.sh` to `tests/BUILD`.

Additionally, we should split up the test data directory to be a
one-directory-per test so that it's clear what test has ownership over which
subset of the files.

As a side benefit, we will end up re-running fewer tests if data files are
grouped properly and only a few files are modified, because Bazel keeps track of
the dependency tree across runs.