Remove `regen-test` from Makefile and Travis CI
We don't really need this test, since it will be generated and then validated
with the same `autogen` script: it should always pass.

This would only be useful in the course of local development when updating the
autogen tool itself, e.g., adding support for new filetypes or changing the
behavior of a flag.